aaaaaaaaaaaaaaaaaaaa

+++ Portal/app/Http/Controllers/UserController.php
@@ -23,6 +23,7 @@ class UserController extends Controller
 
         if ($request->hasFile('pfp')) {
             $user = User::findOrFail(Auth::id());
+            if($user->profilepic != "ProfilePictures/defaultpfp.jpg")
             Storage::disk('public')->delete($user->profilepic);
             $pfp = Storage::disk('public')->put('ProfilePictures',$request->pfp);
             $user->update(['profilepic' => $pfp]);
